+++ src/main/java/wooteco/subway/domain/Section.java
@@ -5,7 +5,6 @@ import java.util.Objects;
 public class Section {
 
     private final Long id;
-//    private final long lineId;
     private final Line line;
     private final Station upStation;
     private final Station downStation;
