+++ src/main/java/wooteco/subway/service/PathService.java
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.section.SectionDao;
 import wooteco.subway.dao.station.StationDao;
-import wooteco.subway.domain.Path;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
 import wooteco.subway.domain.strategy.FindPathStrategy;
@@ -19,7 +18,8 @@ public class PathService {
     private final StationDao stationDao;
     private final FindPathStrategy findPathStrategy;
 
-    public PathService(final SectionDao sectionDao, final StationDao stationDao, final FindPathStrategy findPathStrategy) {
+    public PathService(final SectionDao sectionDao, final StationDao stationDao,
+                       final FindPathStrategy findPathStrategy) {
         this.sectionDao = sectionDao;
         this.stationDao = stationDao;
         this.findPathStrategy = findPathStrategy;
