+++ src/test/java/wooteco/subway/acceptance/PathFindAcceptanceTest.java
@@ -1,14 +1,10 @@
 package wooteco.subway.acceptance;
 
 import static io.restassured.RestAssured.given;
-import static wooteco.subway.acceptance.LineAcceptanceTest.postLines;
-import static wooteco.subway.acceptance.SectionAcceptanceTest.postSections;
-import static wooteco.subway.acceptance.StationAcceptanceTest.postStations;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import wooteco.subway.dto.line.LineSaveRequest;
 import wooteco.subway.dto.path.PathFindRequest;
 import wooteco.subway.dto.section.SectionSaveRequest;
