+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -9,24 +9,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import wooteco.subway.dto.station.StationResponse;
 import wooteco.subway.dto.station.StationSaveRequest;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
-
-    public static Long postStations(final StationSaveRequest stationSaveRequest) {
-        return RestAssured.given().log().all()
-                .body(stationSaveRequest)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/stations")
-                .then().log().all()
-                .extract()
-                .response()
-                .as(StationResponse.class)
-                .getId();
-    }
 
     @DisplayName("지하철역을 생성한다.")
     @Test
